Move movie to own component folder, update import
